requirements.txt
openai
faiss-cpu
python-dotenv
PyPDF2
langchain
